requirements.txt
fastapi
uvicorn
pydantic-settings
google-genai
chromadb
sentence-transformers
pymupdf
python-dotenv